--- a/.gitignore
+++ b/.gitignore
@@ -3,4 +3,5 @@ node_modules
 /lib
 coverage
 .vscode
+.idea
 *.log