--- a/data-type-sans-outliers-v16.0.0.svg
+++ b/data-type-sans-outliers-v16.0.0.svg
@@ -1,0 +1,91 @@
+<svg xmlns="http://www.w3.org/2000/svg" class="marks" width="1001" height="820">
+    <path fill="#fff" d="M0 0h1001v820H0z"/>
+    <g class="mark-group role-frame root" aria-roledescription="group mark container" fill="none" stroke-miterlimit="10">
+        <path class="background" aria-hidden="true" d="M116.5 11.5h880v600h-880z" stroke="#ddd"/>
+        <g class="mark-group role-axis" aria-hidden="true">
+            <path class="background" aria-hidden="true" d="M116.5 11.5z" pointer-events="none"/>
+            <g class="mark-rule role-axis-grid" pointer-events="none" stroke="#ddd">
+                <path d="M116.5 611.5h880M116.5 551.5h880M116.5 491.5h880M116.5 431.5h880M116.5 371.5h880M116.5 311.5h880M116.5 251.5h880M116.5 191.5h880M116.5 131.5h880M116.5 71.5h880M116.5 11.5h880"/>
+            </g>
+        </g>
+        <g class="mark-group role-axis" aria-roledescription="axis" aria-label="X-axis titled 'module' for a discrete scale with 23 values: @mojotech/json-type-validation, @typeofweb/schema, bueno, class-transformer-validator-async, class-transformer-validator-sync, ending with zod">
+            <path class="background" aria-hidden="true" d="M116.5 611.5z" pointer-events="none"/>
+            <g class="mark-rule role-axis-tick" pointer-events="none" stroke="#888">
+                <path d="M135.5 611.5v5M173.5 611.5v5M211.5 611.5v5M249.5 611.5v5M288.5 611.5v5M326.5 611.5v5M364.5 611.5v5M402.5 611.5v5M441.5 611.5v5M479.5 611.5v5M517.5 611.5v5M556.5 611.5v5M594.5 611.5v5M632.5 611.5v5M670.5 611.5v5M709.5 611.5v5M747.5 611.5v5M785.5 611.5v5M823.5 611.5v5M862.5 611.5v5M900.5 611.5v5M938.5 611.5v5M976.5 611.5v5"/>
+            </g>
+            <g class="mark-text role-axis-label" pointer-events="none" font-family="sans-serif" font-size="14" fill="#000">
+                <text transform="rotate(90 -243.685 374.815)">@mojotech/json-…</text>
+                <text transform="rotate(90 -224.554 393.946)">@typeofweb/sche…</text>
+                <text transform="rotate(90 -205.424 413.076)">bueno</text>
+                <text transform="rotate(90 -186.293 432.207)">class-transform…</text>
+                <text transform="rotate(90 -167.163 451.337)">class-transform…</text>
+                <text transform="rotate(90 -148.033 470.467)">computed-types</text>
+                <text transform="rotate(90 -128.902 489.598)">decoders</text>
+                <text transform="rotate(90 -109.772 508.728)">io-ts</text>
+                <text transform="rotate(90 -90.641 527.859)">jointz</text>
+                <text transform="rotate(90 -71.51 546.99)">json-decoder</text>
+                <text transform="rotate(90 -52.38 566.12)">milankinen/zod</text>
+                <text transform="rotate(90 -33.25 585.25)">myzod</text>
+                <text transform="rotate(90 -14.12 604.38)">purify-ts</text>
+                <text transform="rotate(90 5.01 623.51)">rulr</text>
+                <text transform="rotate(90 24.141 642.641)">runtypes</text>
+                <text transform="rotate(90 43.272 661.772)">simple-runtypes</text>
+                <text transform="rotate(90 62.402 680.902)">superstruct</text>
+                <text transform="rotate(90 81.533 700.033)">toi</text>
+                <text transform="rotate(90 100.663 719.163)">ts-interface-ch…</text>
+                <text transform="rotate(90 119.793 738.293)">ts-utils</text>
+                <text transform="rotate(90 138.924 757.424)">valita</text>
+                <text transform="rotate(90 158.054 776.554)">yup</text>
+                <text transform="rotate(90 177.185 795.685)">zod</text>
+            </g>
+            <path stroke="#888" d="M116.5 611.5h880" class="mark-rule role-axis-domain" pointer-events="none"/>
+            <text text-anchor="middle" transform="translate(556.5 812.5)" font-family="sans-serif" font-size="14" fill="#000" class="mark-text role-axis-title" pointer-events="none">module</text>
+        </g>
+        <g class="mark-group role-axis" aria-roledescription="axis" aria-label="Y-axis titled 'operations / sec' for a linear scale with values from 0 to 5,000,000">
+            <path class="background" aria-hidden="true" d="M116.5 11.5z" pointer-events="none"/>
+            <g class="mark-rule role-axis-tick" pointer-events="none" stroke="#888">
+                <path d="M116.5 611.5h-5M116.5 551.5h-5M116.5 491.5h-5M116.5 431.5h-5M116.5 371.5h-5M116.5 311.5h-5M116.5 251.5h-5M116.5 191.5h-5M116.5 131.5h-5M116.5 71.5h-5M116.5 11.5h-5"/>
+            </g>
+            <g class="mark-text role-axis-label" pointer-events="none" text-anchor="end" font-family="sans-serif" font-size="12" fill="#000">
+                <text transform="translate(109.5 615.5)">0</text>
+                <text transform="translate(109.5 555.5)">500,000</text>
+                <text transform="translate(109.5 495.5)">1,000,000</text>
+                <text transform="translate(109.5 435.5)">1,500,000</text>
+                <text transform="translate(109.5 375.5)">2,000,000</text>
+                <text transform="translate(109.5 315.5)">2,500,000</text>
+                <text transform="translate(109.5 255.5)">3,000,000</text>
+                <text transform="translate(109.5 195.5)">3,500,000</text>
+                <text transform="translate(109.5 135.5)">4,000,000</text>
+                <text transform="translate(109.5 75.5)">4,500,000</text>
+                <text transform="translate(109.5 15.5)">5,000,000</text>
+            </g>
+            <path stroke="#888" d="M116.5 611.5v-600" class="mark-rule role-axis-domain" pointer-events="none"/>
+            <text text-anchor="middle" transform="rotate(-90 164 147.5)" font-family="sans-serif" font-size="14" fill="#000" class="mark-text role-axis-title" pointer-events="none">operations / sec</text>
+        </g>
+        <g class="mark-rect role-mark marks" aria-roledescription="rect mark container" fill="#4c78a8">
+            <path aria-label="operations / sec: 103351; module: bueno" aria-roledescription="bar" d="M194.435 598.598h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 54053; module: class-transformer-validator-async" aria-roledescription="bar" d="M232.696 604.514h34.434V611h-34.434z"/>
+            <path aria-label="operations / sec: 66729; module: class-transformer-validator-sync" aria-roledescription="bar" d="M270.957 602.993h34.434V611h-34.434z"/>
+            <path aria-label="operations / sec: 1302277; module: computed-types" aria-roledescription="bar" d="M309.217 454.727h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 328543; module: decoders" aria-roledescription="bar" d="M347.478 571.575h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 1924017; module: io-ts" aria-roledescription="bar" d="M385.74 380.118h34.434V611h-34.435z"/>
+            <path aria-label="operations / sec: 2303597; module: jointz" aria-roledescription="bar" d="M424 334.568h34.435V611H424z"/>
+            <path aria-label="operations / sec: 1205792; module: json-decoder" aria-roledescription="bar" d="M462.26 466.305h34.436V611H462.26z"/>
+            <path aria-label="operations / sec: 1801597; module: @mojotech/json-type-validation" aria-roledescription="bar" d="M117.913 394.808h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 2115310; module: myzod" aria-roledescription="bar" d="M538.783 357.163h34.434V611h-34.434z"/>
+            <path aria-label="operations / sec: 1023057; module: purify-ts" aria-roledescription="bar" d="M577.043 488.233h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 1242731; module: rulr" aria-roledescription="bar" d="M615.304 461.872h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 161748; module: runtypes" aria-roledescription="bar" d="M653.565 591.59H688V611h-34.435z"/>
+            <path aria-label="operations / sec: 3077120; module: simple-runtypes" aria-roledescription="bar" d="M691.826 241.746h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 152440; module: superstruct" aria-roledescription="bar" d="M730.087 592.707h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 874245; module: toi" aria-roledescription="bar" d="M768.348 506.09h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 64927; module: ts-utils" aria-roledescription="bar" d="M844.87 603.209h34.434V611H844.87z"/>
+            <path aria-label="operations / sec: 1320077; module: @typeofweb/schema" aria-roledescription="bar" d="M156.174 452.59h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 4634932; module: valita" aria-roledescription="bar" d="M883.13 54.808h34.435V611H883.13z"/>
+            <path aria-label="operations / sec: 42782; module: yup" aria-roledescription="bar" d="M921.391 605.866h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 2981; module: zod" aria-roledescription="bar" d="M959.652 610.642h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 469237; module: milankinen/zod" aria-roledescription="bar" d="M500.522 554.692h34.435V611h-34.435z"/>
+            <path aria-label="operations / sec: 3340228; module: ts-interface-checker" aria-roledescription="bar" d="M806.609 210.173h34.434V611H806.61z"/>
+        </g>
+    </g>
+</svg>