Hello and first thanks for this project! I've hoped to find a validation library that focuses on data validation, has simple and intuitive API and can infer TS types correctly from the schemas. When I encountered `zod`, I knew that it was exactly what I had been searching for. So I happily started using the library but soon I unfortunately realized that there were some performance issues that made the validations run very slow and due these issues I had to drop the library from production. I felt very sad about it. 😢 So here is my contribution that'd hopefully help to fix some of these issues in future.

tl;dr; this PR introduces some changes that bring **100-200x speedup** compared to the current `master` 

### Baseline

There is already some prior research done about the performance, such as #292 and #459 which have found that excessive `ZodError` allocation is the main cause for poor performance and one fix would be to remove the `Error` base class (and hence remove the costly stacktrace allocation). And if I've understood correctly, the main concern is that the stacktrace could be relevant for the library uses (and to be honest, I share this concern as well).

So I took a look at the codebase and (like @reinismu [said](https://github.com/colinhacks/zod/pull/292#discussion_r567867700)) it heavily overuses `try/catch`. The codebase also implements `PseudoPromise` abstraction that allows to support async refinements. If we (mentally) unroll the structure, the code flow looks like the following:

```js
// level 1
try {
  const err1 = new ZodError()
  if (invalid(level1)) throw err1
  // level 2
  try {
    const err2 = new ZodError()
    if (invalid(level2)) throw err2
    // level 3 
    try {
      const err3 = new ZodError()
      if (invalid(level3)) throw err3
      // level ...
      try {
        // ...
      } catch (err) {
        // ...
      }
    } catch (err) {
      err2.addIssues(err.issues)
      throw err2
    }
  } catch (err) {
    err1.addIssues(err.issues)
    throw err1
  }
} catch (err) {
  baseErr.addIssues(err.issues)
  throw baseErr
}
```

Like you can see, the code allocates quite many errors. It also accumulates issues per level which is a lot of extra work. In addition this flow is orchestrated by `PseudoPromise` abstraction that also adds some extra allocation and calculation to the parsing.

At this point, I ran the (excellent!) benchmarks introduced in #431 by @scotttrinh and got the following results on my '20 MBP (Node 16):
<details>
  <summary> ** Baseline (~50k ops/s) ** </summary>
  
```
z.string: empty string x 69,126 ops/sec ±1.71% (86 runs sampled)
z.string: short string x 66,917 ops/sec ±4.43% (85 runs sampled)
z.string: long string x 69,844 ops/sec ±4.63% (87 runs sampled)
z.string: invalid: null x 51,296 ops/sec ±4.51% (82 runs sampled)
z.string: manual parser: long x 878,697,948 ops/sec ±1.60% (85 runs sampled)
z.object: empty: valid x 31,908 ops/sec ±8.31% (81 runs sampled)
z.object: empty: valid: extra keys x 35,451 ops/sec ±6.17% (84 runs sampled)
z.object: empty: invalid: null x 56,126 ops/sec ±1.50% (89 runs sampled)
z.object: short: valid x 20,422 ops/sec ±6.28% (84 runs sampled)
z.object: short: valid: extra keys x 19,344 ops/sec ±6.31% (79 runs sampled)
z.object: short: invalid: null x 50,725 ops/sec ±1.61% (82 runs sampled)
z.object: long: valid x 10,492 ops/sec ±3.93% (80 runs sampled)
z.object: long: valid: extra keys x 9,877 ops/sec ±7.00% (80 runs sampled)
z.object: long: invalid: null x 51,332 ops/sec ±1.65% (88 runs sampled)
```

</details>


### 1. Optimize flow control

 
So what I did is that I changed the execution flow into more lightweight variant which allows better composition of parsing functions and minimizes the required allocations, in pseudo:

```js 
const level1Valid = valid(level1Data)
const level2Valid = level1Valid && valid(level2Data)
const level3Valid = level2Valid && valid(level3Data)
const levelNParsed = level3Valid && valid(levelNData) ? makeResult(levelNData) : INVALID
const level3Parsed = levelNParsed === INVALID ? INVALID : makeResult(level3Data, levelNParsed)
const level2Parsed = level3Parsed === INVALID ? INVALID : makeResult(level2Data, level3Parsed)
const level1Parsed = level2Parsed === INVALID ? INVALID : makeResult(level1Data, level3Parsed)
```
Like in the current implementation, the parsing context is passed as a parameter to `_parse` but the difference is that errors are not indicated by throwing and catching them. Instead, `_parse` returns a result object of `ParseResult = INVALID | OK(parsed)`. If any parsing errors occur, issues are stored directly into the context: we known that if parsing fails, the call stack will eventually be returning to the top level `parse` (or `safeParse`) where we can read all the issues and **then** throw an error with all of those issues. This means that **there are no breaking changes to the public API** but the difference in performance is huge.

Of course the implementation must also support async refinements so the parse result type is actually `ParseResult = INVALID | OK(parsed) | ASYNC(INVALID | OK(parsed))` and parsers must implement the async case as well. Luckily the number of those cases is relatively low (basically collections like object, array, tuple, record, set, list) so I could add specialized performance-optimized implementations for each case. I ran benchmarks again and the different is significant. And the best part is that there were absolutely **no changes to the public API** - all tests are passing without any modifications on them!

<details>
  <summary> ** Optimized flow (~5M ops/s) ** </summary>
  
```
z.string: empty string x 13,120,643 ops/sec ±1.27% (94 runs sampled)
z.string: short string x 12,989,836 ops/sec ±1.45% (92 runs sampled)
z.string: long string x 12,788,053 ops/sec ±1.21% (91 runs sampled)
z.string: invalid: null x 118,915 ops/sec ±1.32% (81 runs sampled)
z.string: manual parser: long x 876,793,503 ops/sec ±1.38% (82 runs sampled)
z.object: empty: valid x 5,437,250 ops/sec ±4.61% (88 runs sampled)
z.object: empty: valid: extra keys x 5,086,771 ops/sec ±1.36% (88 runs sampled)
z.object: empty: invalid: null x 127,522 ops/sec ±3.38% (90 runs sampled)
z.object: short: valid x 3,103,337 ops/sec ±1.53% (93 runs sampled)
z.object: short: valid: extra keys x 3,223,728 ops/sec ±1.64% (90 runs sampled)
z.object: short: invalid: null x 126,815 ops/sec ±4.66% (91 runs sampled)
z.object: long: valid x 1,482,835 ops/sec ±1.32% (94 runs sampled)
z.object: long: valid: extra keys x 1,411,238 ops/sec ±1.22% (91 runs sampled)
z.object: long: invalid: null x 120,033 ops/sec ±3.74% (87 runs sampled)
```

</details>

### 2. Add some extra optimizations

In addition to the flow, I also did some minor optimizations:

* https://github.com/milankinen/zod/commit/341e9ddc46dbf12f9860c8aa7c5eb177e9bc27ad - change context path updates to use persistent O(1) list instead of O(n) array copy
* https://github.com/milankinen/zod/commit/423c9a97f14fec5ce0b66d5f74378802161a9c97 - cache object shape information over subsequent parse events
* https://github.com/milankinen/zod/commit/89088c2570230d9dd17e57c5e795b9babba69c57 - switch-case parsed type calculation

With these smaller local optimizations, I could squeeze some extra cycles:

<details>
  <summary> ** Optimized flow + local optimizations (~7,5M ops/s) ** </summary>
  
```
z.string: empty string x 11,050,585 ops/sec ±1.34% (89 runs sampled)
z.string: short string x 10,931,303 ops/sec ±1.58% (89 runs sampled)
z.string: long string x 11,056,030 ops/sec ±1.43% (94 runs sampled)
z.string: invalid: null x 118,733 ops/sec ±1.41% (90 runs sampled)
z.string: manual parser: long x 864,371,553 ops/sec ±1.41% (82 runs sampled)
z.object: empty: valid x 7,120,299 ops/sec ±4.26% (88 runs sampled)
z.object: empty: valid: extra keys x 7,585,016 ops/sec ±1.28% (93 runs sampled)
z.object: empty: invalid: null x 129,548 ops/sec ±3.68% (89 runs sampled)
z.object: short: valid x 4,777,457 ops/sec ±1.35% (94 runs sampled)
z.object: short: valid: extra keys x 5,014,443 ops/sec ±4.73% (90 runs sampled)
z.object: short: invalid: null x 127,899 ops/sec ±3.33% (89 runs sampled)
z.object: long: valid x 2,173,238 ops/sec ±1.34% (94 runs sampled)
z.object: long: valid: extra keys x 2,125,664 ops/sec ±1.29% (91 runs sampled)
z.object: long: invalid: null x 117,491 ops/sec ±3.75% (81 runs sampled)
```

</details>

I also ran the https://github.com/moltar/typescript-runtime-type-benchmarks with current master (`zod` in the graphics) and optimizations (`milankinen/zod` in the graphics) and got the following results:

![Som graphics](https://raw.githubusercontent.com/milankinen/zod/performance-improvements/data-type-sans-outliers-v16.0.0.png)

<details>
  <summary> Full onsole output </summary>
  
```
Running "data-type-sans-outliers" suite...
 decoders:
    328 543 ops/s, ±1.64%     | 92.91% slower

  io-ts:
    1 924 017 ops/s, ±1.39%   | 58.49% slower

  jointz:
    2 303 597 ops/s, ±1.18%   | 50.3% slower

  json-decoder:
    1 205 792 ops/s, ±1.44%   | 73.98% slower

  @mojotech/json-type-validation:
    1 801 597 ops/s, ±1.40%   | 61.13% slower

  myzod:
    2 115 310 ops/s, ±1.29%   | 54.36% slower

  purify-ts:
    1 023 057 ops/s, ±1.73%   | 77.93% slower

  rulr:
    1 242 731 ops/s, ±1.36%   | 73.19% slower

  runtypes:
    161 748 ops/s, ±1.66%     | 96.51% slower

  simple-runtypes:
    3 077 120 ops/s, ±1.23%   | 33.61% slower

  superstruct:
    152 440 ops/s, ±1.81%     | 96.71% slower

  toi:
    874 245 ops/s, ±1.41%     | 81.14% slower

  ts-utils:
    64 927 ops/s, ±60.92%      | 98.6% slower

  @typeofweb/schema:
    1 320 077 ops/s, ±1.54%   | 71.52% slower

  valita:
    4 634 932 ops/s, ±1.12%   | fastest

  yup:
    42 782 ops/s, ±1.14%      | 99.08% slower

  zod:
    2 981 ops/s, ±17.43%       | slowest, 99.94% slower

  milankinen/zod:
    469 237 ops/s, ±41.76%     | 89.88% slower

  ts-interface-checker:
    3 340 228 ops/s, ±1.29%   | 27.93% slower

Finished 23 cases!
  Fastest: valita
  Slowest: zod

```

</details>

### Conclusions

I'm very happy that I found a way to significantly improve the performance while maintaining this excellent API. Honestly I don't believe or expect that this PR will be merged to the upstream as it is but I at least hope that it'll give some pointers for the future improvements. I really like `zod` as a library and hope that some day the performance gets fixed so that it won't prevent me from using this in production as well. :slightly_smiling_face:

Cheers!