@colinhacks 2021-06-14
PRs like this remind me why I love open source. Truly glorious stuff Matti!

I've been looking through the code and I think I understand the new execution system. I like it! Getting rid of all those spaghetti code PseudoPromise methods is so satisfying.

For grins, I re-ran the benchmarks after removing the `Error` base class, and saw nearly 3x performance improvements on all invalid results. Which is a pretty compelling reason to remove it...

```
z.string: empty string x 9,702,190 ops/sec ±0.89% (87 runs sampled)
z.string: short string x 9,628,797 ops/sec ±0.84% (92 runs sampled)
z.string: long string x 9,472,395 ops/sec ±2.01% (88 runs sampled)
z.string: invalid: null x 419,983 ops/sec ±1.34% (87 runs sampled)
z.string: manual parser: long x 973,337,378 ops/sec ±0.75% (96 runs sampled)
z.object: empty: valid x 7,065,026 ops/sec ±1.18% (87 runs sampled)
z.object: empty: valid: extra keys x 8,204,405 ops/sec ±0.94% (92 runs sampled)
z.object: empty: invalid: null x 537,947 ops/sec ±1.76% (86 runs sampled)
z.object: short: valid x 4,047,468 ops/sec ±0.96% (89 runs sampled)
z.object: short: valid: extra keys x 4,494,936 ops/sec ±1.06% (86 runs sampled)
z.object: short: invalid: null x 554,633 ops/sec ±1.06% (92 runs sampled)
z.object: long: valid x 2,041,231 ops/sec ±1.07% (86 runs sampled)
z.object: long: valid: extra keys x 2,031,685 ops/sec ±1.13% (88 runs sampled)
z.object: long: invalid: null x 535,857 ops/sec ±2.65% (82 runs sampled)
```

You say you don't expect that this PR will be merged...but I can't see any reason why not. Seems like a definitive improvement in terms of code understandability, perf, and elegance. Am I missing anything?

I think combining this with some of the changes by @scotttrinh will be the best of all worlds: changing the format of `ZodInvalidUnionIssue` and eliminating the Error subclass for ZodError (cc @lilred). Scott what do you think?