@milankinen 2021-06-14
>You say you don't expect that this PR will be merged...but I can't see any reason why not. Seems like a definitive improvement in terms of code understandability, perf, and elegance. Am I missing anything?

Well, this PR introduces quite big changes to the internals which can be an issue sometimes. 😅 But I'm glad that you approve those changes, technically there should be no reason not to merge this (except the conflicts that must be resolved first). 🙂 This PR is forked from #431 so perhaps that could be merged first as a separate PR?

>For grins, I re-ran the benchmarks after removing the Error base class, and saw nearly 3x performance improvements on all invalid results. Which is a pretty compelling reason to remove it...

Yeah, removing the `Error` class can improve the performance for invalid cases quite a lot so if bumping the major version does not scare, then we should definitely go for it!

Another, a bit more conservative option could be to keep `Error` sub-classing in `.parse` and `.parseAsync` so that the thrown error has a stacktrace, but return `{success: false, issues: ZodIssue[])` from `safeParse` and `spa`? That said, both approaches are not backward compatible so they'd require a major version bump. 

>I think combining this with some of the changes by @scotttrinh will be the best of all worlds: changing the format of ZodInvalidUnionIssue and eliminating the Error subclass for ZodError (cc @lilred). Scott what do you think?

I definitely agree with this 👍 Storing error instances inside another error won't give any benefits.