@scotttrinh 2021-06-14
> I think combining this with some of the changes by @scotttrinh will be the best of all worlds: changing the format of `ZodInvalidUnionIssue` and eliminating the Error subclass for ZodError (cc @lilred). Scott what do you think?

I don't think we should wait on my opinion here, I was mostly working off existing work! I'll have time in the next few hours to review and it sounds like we will need to rebase or deconflict this on the latest, so I'm happy to help with that, too.