@milankinen 2021-06-14
@scotttrinh thanks for the review comments! I committed the suggested changes. There are still some conflicting files because this branch is built on top of #431 which is a little bit behind the current master. What do you think is the best option? I could either:

1. Rebase this whole branch with the benchmark commits onto current master
2. Rebase only my commits and leave benchmark branch to be merged separately

Also one other thing I noticed when I was running the tests with the new implementation was that [this test](https://github.com/colinhacks/zod/blob/master/src/__tests__/async-parsing.test.ts#L364-L387) seemed to have an expectation that was conflicting with the description. The description says that "early async failure prevents follow-up refinement checks". However, the expected `count` value is 2, indicating that both refinements are run, even though the first one returns `false`. I didn't want to break the existing behaviour so I removed the refinement prevention but what is actually the expected behaviour?