@reinismu 2021-06-14
Nice work!
Still need some work to reach https://github.com/badrap/valita level, but at least it shouldn't be the bottleneck of the whole application after this :)