@milankinen 2021-06-14
Thanks! Valita is indeed a nice piece of software. Of course running these kind of microbenchmarks is a bit dangerous because benchmark setup might favour some libraries. For example valita seems to be built built on top of assumption that validated object does not contain any extra keys and if that happens, it falls back to a slower execution path (its still pretty damn fast though 😅!). In `zod` extra keys do not matter (if running with default .strip mode).

If we add some extra keys to the data, we can se that amongst "popular" libs we're almost in par 😄 
```
> typescript-runtime-type-benchmarks@1.0.0 start
> ts-node index.ts

Running "data-type-sans-outliers" suite...
Progress: 100%

  io-ts:
    1 562 873 ops/s, ±31.23%   | 11.62% slower

  runtypes:
    166 728 ops/s, ±0.70%     | 90.57% slower

  valita:
    1 768 350 ops/s, ±0.72%   | fastest

  zod:
    3 458 ops/s, ±14.52%       | slowest, 99.8% slower

  milankinen/zod:
    801 172 ops/s, ±1.18%     | 54.69% slower

Finished 5 cases!
  Fastest: valita
  Slowest: zod
```