@scotttrinh 2021-06-15
@milankinen 

> Also one other thing I noticed when I was running the tests with the new implementation was that this test seemed to have an expectation that was conflicting with the description. The description says that "early async failure prevents follow-up refinement checks". However, the expected count value is 2, indicating that both refinements are run, even though the first one returns false. I didn't want to break the existing behaviour so I removed the refinement prevention but what is actually the expected behaviour?

I'd say even if the original _intent_ was to prevent the refinement, the fact that the current version does not prevent it (given the tests) means that's the behavior to preserve. If we're collecting possible breaking changes around these updates here, maybe it's worth considering that more holistically, but I imagine we want to minimize breakage. Definitely a @colinhacks call.

> There are still some conflicting files because this branch is built on top of #431 which is a little bit behind the current master. What do you think is the best option?

I'm curious about this question as well. In my opinion, I'd rather see us concentrate on landing your changes here while incorporating some of the additional changes to `ZodError` that the other PR introduces, but I want to get some marching orders from Colin around this since it seems like we're all in a good position to take on the work.